added gradient method
